requirement.txt
pandas
numpy
scipy
scikit-learn
sentence-transformers
torch
google-generativeai
python-dotenv